Fixed Rounding Error Better

+++ src/Account.java
@@ -1,20 +1,21 @@
+import java.math.BigDecimal;
 
 public class Account {
 
 	Long id;
-	double balence;
+	BigDecimal balence;
 	
 	public Account (Long id) {
 		this.id = id;
-		this.balence = 0.0;
+		this.balence = BigDecimal.valueOf(0.0);
 	}
 	
 	public Account (Long id, double balence) {
 		this.id = id;
-		this.balence = balence;
+		this.balence = BigDecimal.valueOf(balence);
 	}
 	
-	public void setBalence(double balence) {
+	public void setBalence(BigDecimal balence) {
 		this.balence = balence;
 	}
